requirements.txt
# Core dependencies
pandas
numpy
torch
openai
sentence-transformers
openpyxl
requests
selenium
imagehash
matplotlib
scipy
neo4j>=5.0

# Optional tools for development, testing, and exploration
jupyter
notebook
ipykernel
ipywidgets
tqdm